+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.5.0 (TBD)
 
+* New Cursor methods: has_children(), to_document(), insert_document()
+
 # 0.4.0 (2025-11-02)
 
 * SyncCursor is now properly Send&Sync.